Crate quoted_string

This crate provides utilities to handle quoted strings like such appearing in Media Types (both MIME (i.e. Mail) and HTTP). As there are many small but significant differences in different specifications this crate does not provide a specific implementation. Instead a QuotedStringSpec trait is exposed. Implementing it (on zero-sized structs) should allow the usage with any quoted-string specification.

Modules

error
module containing all errors

spec
This module contains types for specifying what kind of quoted string is used

test_utils
provides an example implementation of quoted string spec's

Structs

ContentChars
A iterator over chars of the content represented by the quoted strings (PartialEq<&str>)

Parsed
the result of successfully parsing a quoted string

Traits

AsciiCaseInsensitiveEq
Analogous to PartialEq, but with ascii case insensitive equality

Functions

parse
parse a quoted string starting at the begin of input but possible ending earlier

quote
quotes the input string returning the quoted string

quote_if_needed
quotes the input string if needed

strip_dquotes
strips quotes if they exists

to_content
converts a quoted string into it's content

validate
validates if input is a valid quoted-string