Module quoted_string::spec

This module contains types for specifying what kind of quoted string is used

Structs

AsciiWordValidator
Allows unquoted text containing only _ | a..z | A..Z | 0..9

PartialCodePoint
A type which represents part of a utf-8 code point

ScanAutomaton

Enums

QuotingClass
Represents if a char can be contained in a quoted string and if it needs escapeing

State
State used when parsing a quoted string

Traits

GeneralQSSpec
type to specify the quoting classifier and parsing implementation

ParsingImpl
This normally zero sized type provides functions for parsing a quoted string

QuotingClassifier
Type to provide a quoting classification method.

WithoutQuotingValidator
Used to validate if a string is valid without beeing quoted.